Shard exact task-shape contracts by language
